fixing an overexited cast

+++ libafl/src/events/launcher.rs
@@ -459,7 +459,7 @@ where
                             ));
 
                             let client_id = ClientDescription::new(
-                                (core_i * self.overcommit + overcommit_i) as u64,
+                                (core_i * self.overcommit + overcommit_i),
                                 overcommit_i,
                                 CoreId(core_i),
                             );
